requirements.txt
# Web框架和基础依赖
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 数据库
motor>=3.0.0
pymongo>=4.0.0
beanie>=1.18.0

# 异步和网络
aiohttp>=3.8.0
websockets>=10.0
requests-oauthlib>=1.3.0

# 科学计算和数据处理
numpy>=1.21.0
shapely>=2.0.0
matplotlib>=3.4.0
osmnx>=1.3.0
networkx>=3.0
gymnasium>=0.29.0

# 计算机视觉
opencv-python>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# AI代理框架
camel-ai>=0.1.0
stable-baselines3>=2.0.0
langchain>=0.0.340
langchain-community>=0.0.20

# 工具和配置
python-dotenv>=0.19.0
loguru>=0.7.0